refactor(mdms-flink): move read/write utilities to own package
