+++ mdms/mdms-flink/src/main/java/de/hpi/isg/mdms/flink/serializer/DVOFlinkSerializer.java
@@ -8,7 +8,7 @@ import org.apache.flink.api.java.typeutils.TupleTypeInfo;
 
 import de.hpi.isg.mdms.domain.RDBMSMetadataStore;
 import de.hpi.isg.mdms.domain.constraints.DistinctValueOverlap;
-import de.hpi.isg.mdms.flink.SqLiteJDBCInputFormat;
+import de.hpi.isg.mdms.flink.readwrite.SqLiteJDBCInputFormat;
 import de.hpi.isg.mdms.model.MetadataStore;
 import de.hpi.isg.mdms.model.constraints.ConstraintCollection;
